requirements.txt
Fabric
boto
click
pyaml
